renamed images and updated their references
