+++ mobile/assets/mascot_waving.png
